git tehlek

+++ src/main/java/tn/esprit/spring/wecare/entities/BestAndWorstPost.java
@@ -2,5 +2,6 @@ package tn.esprit.spring.wecare.entities;
 
 public interface BestAndWorstPost {
 	public String getname();
-	public double getocc();
+	public int getocc();
 }
+
